# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Smart City Digital Citizen Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /taxes/property:
    get:
      operationId: listPropertyTaxes
      tags: [taxes]
      summary: List property tax bills
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, paid, overdue]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Property tax bills
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyTaxListResponse"
  /taxes/property/{id}:
    get:
      operationId: getPropertyTax
      tags: [taxes]
      summary: Get property tax bill by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Property tax bill details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyTax"
  /taxes/property/{id}/pay:
    post:
      operationId: payPropertyTax
      tags: [taxes]
      summary: Pay property tax bill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "200":
          description: Payment receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentReceipt"
  /taxes/water:
    get:
      operationId: listWaterTaxes
      tags: [taxes]
      summary: List water tax bills
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, paid, overdue]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Water tax bills
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WaterTaxListResponse"
  /taxes/water/{id}:
    get:
      operationId: getWaterTax
      tags: [taxes]
      summary: Get water tax bill by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Water tax bill details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WaterTax"
  /taxes/water/{id}/pay:
    post:
      operationId: payWaterTax
      tags: [taxes]
      summary: Pay water tax bill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "200":
          description: Payment receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentReceipt"
...
components:
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [total, page, limit, totalPages]
...
    PropertyTax:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        propertyId:
          type: string
        propertyAddress:
          type: string
        propertyType:
          type: string
          enum: [residential, commercial, industrial]
        assessedValue:
          type: number
        taxRate:
          type: number
        taxAmount:
          type: number
        penaltyAmount:
          type: number
        totalDue:
          type: number
        dueDate:
          type: string
          format: date
        status:
          type: string
          enum: [pending, paid, overdue]
        financialYear:
          type: string
        paidAt:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, userId, propertyId, propertyAddress, propertyType, assessedValue, taxRate, taxAmount, penaltyAmount, totalDue, dueDate, status, financialYear, createdAt]
    PropertyTaxListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/PropertyTax"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
      required: [data, pagination]

    # ─── Water Tax ───
    WaterTax:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        connectionId:
          type: string
        connectionAddress:
          type: string
        meterReading:
          type: number
        previousReading:
          type: number
        unitsConsumed:
          type: number
        ratePerUnit:
          type: number
        taxAmount:
          type: number
        penaltyAmount:
          type: number
        totalDue:
          type: number
        billingPeriod:
          type: string
        dueDate:
          type: string
          format: date
        status:
          type: string
          enum: [pending, paid, overdue]
        paidAt:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, userId, connectionId, connectionAddress, meterReading, previousReading, unitsConsumed, ratePerUnit, taxAmount, penaltyAmount, totalDue, billingPeriod, dueDate, status, createdAt]
    WaterTaxListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/WaterTax"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
      required: [data, pagination]
...
    PaymentInput:
      type: object
      properties:
        paymentMethod:
          type: string
          enum: [credit_card, debit_card, net_banking, upi, wallet]
        cardLast4:
          type: string
      required: [paymentMethod]
    PaymentReceipt:
      type: object
      properties:
        paymentId:
          type: string
        receiptNumber:
          type: string
        amount:
          type: number
        paymentMethod:
          type: string
        status:
          type: string
        paidAt:
          type: string
          format: date-time
      required: [paymentId, receiptNumber, amount, paymentMethod, status, paidAt]
...